HDDS-10934. Refactor TestOzoneRpcClient hierarchy (#6747)
(cherry picked from commit 385c4ec6ca2f20b43477050f918b7a644f7569f5)

 Conflicts:
	hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/ozone/client/rpc/TestSecureOzoneRpcClient.java


HDDS-10934 amendment
